// src/services/cockpit/interfaces/general.ts
export interface Statistics {
  revenues: {
    goal: number;
    result: number;
    percentage: number;
    formattedGoal: string;
    formattedResult: string;
    formattedPercentage: string;
  };
  pog: {
    goal: number;
    result: number;
    percentage: number;
    formattedGoal: string;
    formattedResult: string;
    formattedPercentage: string;
  };
}